Moved more code to bounds
